feat: changed structure of LOC data
Now, the LOC stores a `map<string, map<string, string>` where the first one
is the hash, and the second map has the string in every language by it's
language code.

This is quite similar to the way other libraries (OMI for example)
structure the data. I still have yet to add helper functions.